requirements.txt
streamlit==1.34.0
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
segmentation_models_pytorch==0.3.3  